Split evidence-matrix.ts into outcome-view module + constants
evidence-matrix.ts previously held both the wizard's evidence matrix
data and the dispute outcome view's tri-state field-status logic. This
splits along that natural seam:

  - constants/high-impact-fields.ts    DISPUTE_HIGH_IMPACT_FIELDS +
                                       emptyByProductType helper
  - constants/topical-fields.ts        DISPUTE_TOPICAL_FIELDS
                                       (imports emptyByProductType)
  - constants/fallback-field-labels.ts FALLBACK_EVIDENCE_FIELD_LABELS
  - evidence-field-status.ts           findMatrixLabel, resolveFieldLabel,
                                       collectMatrixKeys, getExpectedFieldStatus
  - evidence-matrix.ts                 wizard data + getMatrixFields only

Test file renamed evidence-matrix.test.ts -> evidence-field-status.test.ts
to match its actual subject.
