+++ client/disputes/new-evidence/constants/high-impact-fields.ts
@@ -1,0 +1,248 @@
+/**
+ * Internal dependencies
+ */
+import type { DisputeReason, ProductType } from 'wcpay/types/disputes';
+
+/**
+ * Build an empty `Record<ProductType, string[]>` for use as a default cell
+ * across reasons in `DISPUTE_HIGH_IMPACT_FIELDS` and `DISPUTE_TOPICAL_FIELDS`.
+ * Returns a fresh object each call so callers can reuse it without sharing
+ * mutable references between cells.
+ *
+ * Imported by `topical-fields.ts` as a sibling.
+ */
+export const emptyByProductType = (): Record< ProductType, string[] > => ( {
+	physical_product: [],
+	digital_product_or_service: [],
+	offline_service: [],
+	event: [],
+	booking_reservation: [],
+	multiple: [],
+	other: [],
+} );
+
+/**
+ * Fields whose presence on a dispute correlates with a higher win rate,
+ * per (reason, product type). Consumed by the Dispute Outcome View to flag
+ * missing high-impact evidence.
+ *
+ * Keys are raw Stripe `dispute.evidence` field names (text and document
+ * fields alike). Cells with an empty array have no data-backed signal
+ * and produce no `expected_missing` markers in the tri-state renderer.
+ *
+ * Auto-populated and unreliable fields are intentionally excluded:
+ *   - `customer_purchase_ip` (Stripe + WooPayments auto-fill on every save)
+ *   - `customer_name`, `customer_email_address`, `billing_address` (Stripe auto-fill)
+ *   - `product_description` (hybrid auto+merchant; placeholder string by
+ *     default — Q6 lift signal is denominator artifact)
+ *   - `uncategorized_file`, `uncategorized_text` (catch-alls; not actionable
+ *     guidance on their own)
+ */
+// eslint-disable-next-line @typescript-eslint/naming-convention -- This is a constant object.
+export const DISPUTE_HIGH_IMPACT_FIELDS: Record<
+	DisputeReason,
+	Record< ProductType, string[] >
+> = {
+	credit_not_processed: {
+		// `customer_signature` (signed delivery proof) is scoped to
+		// physical_product because these disputes commonly take the shape
+		// "I returned the product and never got my refund": proving
+		// delivery corroborates the merchant's defence. The field is
+		// intentionally absent from non-physical cells (no shipping
+		// proof to attach).
+		physical_product: [
+			'customer_signature',
+			'customer_communication',
+			'receipt',
+		],
+		digital_product_or_service: [
+			'customer_communication',
+			'receipt',
+			'refund_refusal_explanation',
+		],
+		offline_service: [
+			'customer_communication',
+			'receipt',
+			'refund_refusal_explanation',
+		],
+		event: [
+			'customer_communication',
+			'receipt',
+			'refund_refusal_explanation',
+		],
+		booking_reservation: [
+			'customer_communication',
+			'receipt',
+			'refund_refusal_explanation',
+		],
+		// `multiple` mirrors `physical_product` as a defensible default
+		// for multi-product orders that may include a physical item.
+		multiple: [ 'customer_signature', 'customer_communication', 'receipt' ],
+		other: [
+			'customer_communication',
+			'receipt',
+			'refund_refusal_explanation',
+		],
+	},
+	duplicate: {
+		physical_product: [
+			'duplicate_charge_explanation',
+			'duplicate_charge_documentation',
+			'shipping_documentation',
+			'receipt',
+		],
+		digital_product_or_service: [
+			'duplicate_charge_explanation',
+			'duplicate_charge_documentation',
+			'receipt',
+		],
+		offline_service: [
+			'duplicate_charge_explanation',
+			'duplicate_charge_documentation',
+			'receipt',
+		],
+		event: [
+			'duplicate_charge_explanation',
+			'duplicate_charge_documentation',
+			'receipt',
+		],
+		booking_reservation: [
+			'duplicate_charge_explanation',
+			'duplicate_charge_documentation',
+			'receipt',
+		],
+		multiple: [
+			'duplicate_charge_explanation',
+			'duplicate_charge_documentation',
+			'shipping_documentation',
+			'receipt',
+		],
+		other: [
+			'duplicate_charge_explanation',
+			'duplicate_charge_documentation',
+			'receipt',
+		],
+	},
+	fraudulent: {
+		// Same lift-based picks across all product types for fraudulent.
+		physical_product: [ 'service_date', 'customer_communication' ],
+		digital_product_or_service: [
+			'service_date',
+			'customer_communication',
+		],
+		offline_service: [ 'service_date', 'customer_communication' ],
+		event: [ 'service_date', 'customer_communication' ],
+		booking_reservation: [ 'service_date', 'customer_communication' ],
+		multiple: [ 'service_date', 'customer_communication' ],
+		other: [ 'service_date', 'customer_communication' ],
+	},
+	general: {
+		physical_product: [ 'receipt', 'customer_communication' ],
+		digital_product_or_service: [ 'receipt', 'customer_communication' ],
+		offline_service: [ 'receipt', 'customer_communication' ],
+		event: [ 'receipt', 'customer_communication' ],
+		booking_reservation: [ 'receipt', 'customer_communication' ],
+		multiple: [ 'receipt', 'customer_communication' ],
+		other: [ 'receipt', 'customer_communication' ],
+	},
+	product_not_received: {
+		physical_product: [
+			'shipping_address',
+			'shipping_tracking_number',
+			'shipping_documentation',
+			'shipping_carrier',
+			'shipping_date',
+		],
+		digital_product_or_service: [
+			'receipt',
+			'customer_communication',
+			'access_activity_log',
+		],
+		offline_service: [ 'receipt', 'customer_communication' ],
+		event: [ 'receipt', 'customer_communication' ],
+		booking_reservation: [ 'receipt', 'customer_communication' ],
+		multiple: [
+			'shipping_address',
+			'shipping_tracking_number',
+			'shipping_documentation',
+			'shipping_carrier',
+			'shipping_date',
+		],
+		other: [ 'receipt', 'customer_communication' ],
+	},
+	product_unacceptable: {
+		physical_product: [
+			'customer_communication',
+			'refund_refusal_explanation',
+			'shipping_documentation',
+		],
+		digital_product_or_service: [
+			'access_activity_log',
+			'customer_communication',
+			'refund_refusal_explanation',
+		],
+		offline_service: [
+			'customer_communication',
+			'refund_refusal_explanation',
+		],
+		// Recommendations for these cells are topical-only (refund_policy,
+		// event/booking documentation). They surface as `optional_missing`
+		// via the existing matrix; no `expected_missing` markers without
+		// data-backed lift.
+		event: [],
+		booking_reservation: [],
+		multiple: [
+			'customer_communication',
+			'refund_refusal_explanation',
+			'shipping_documentation',
+		],
+		other: [],
+	},
+	subscription_canceled: {
+		physical_product: [
+			'cancellation_policy_disclosure',
+			'cancellation_policy',
+			'cancellation_rebuttal',
+		],
+		digital_product_or_service: [
+			'cancellation_policy_disclosure',
+			'cancellation_policy',
+			'cancellation_rebuttal',
+		],
+		offline_service: [
+			'cancellation_policy_disclosure',
+			'cancellation_policy',
+			'cancellation_rebuttal',
+		],
+		event: [
+			'cancellation_policy_disclosure',
+			'cancellation_policy',
+			'cancellation_rebuttal',
+		],
+		booking_reservation: [
+			'cancellation_policy_disclosure',
+			'cancellation_policy',
+			'cancellation_rebuttal',
+		],
+		multiple: [
+			'cancellation_policy_disclosure',
+			'cancellation_policy',
+			'cancellation_rebuttal',
+		],
+		other: [
+			'cancellation_policy_disclosure',
+			'cancellation_policy',
+			'cancellation_rebuttal',
+		],
+	},
+	// No data-backed signal yet: tri-state renders no `expected_missing`
+	// rows for any product type under these reasons.
+	bank_cannot_process: emptyByProductType(),
+	check_returned: emptyByProductType(),
+	customer_initiated: emptyByProductType(),
+	debit_not_authorized: emptyByProductType(),
+	incorrect_account_details: emptyByProductType(),
+	insufficient_funds: emptyByProductType(),
+	noncompliant: emptyByProductType(),
+	unrecognized: emptyByProductType(),
+};
