+++ client/disputes/new-evidence/evidence-matrix.ts
@@ -6,9 +6,8 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import type { RecommendedDocument, EvidenceFieldStatus } from './types';
+import type { RecommendedDocument } from './types';
 import { DOCUMENT_FIELD_KEYS } from './document-field-keys';
-import type { DisputeReason, ProductType } from 'wcpay/types/disputes';
 
 /**
  * Evidence matrix that maps [reason][productType] to recommended document fields.
@@ -1971,585 +1970,4 @@ export const getMatrixFields = (
 
 	// Return the matrix entry for the specific productType, or undefined if not found
 	return evidenceMatrix[ reason ]?.[ productType ];
-};
-
-const emptyByProductType = (): Record< ProductType, string[] > => ( {
-	physical_product: [],
-	digital_product_or_service: [],
-	offline_service: [],
-	event: [],
-	booking_reservation: [],
-	multiple: [],
-	other: [],
-} );
-
-/**
- * Fields whose presence on a dispute correlates with a higher win rate,
- * per (reason, product type). Consumed by the Dispute Outcome View to flag
- * missing high-impact evidence.
- *
- * Keys are raw Stripe `dispute.evidence` field names (text and document
- * fields alike). Cells with an empty array have no data-backed signal
- * and produce no `expected_missing` markers in the tri-state renderer.
- *
- * Auto-populated and unreliable fields are intentionally excluded:
- *   - `customer_purchase_ip` (Stripe + WooPayments auto-fill on every save)
- *   - `customer_name`, `customer_email_address`, `billing_address` (Stripe auto-fill)
- *   - `product_description` (hybrid auto+merchant; placeholder string by
- *     default — Q6 lift signal is denominator artifact)
- *   - `uncategorized_file`, `uncategorized_text` (catch-alls; not actionable
- *     guidance on their own)
- */
-// eslint-disable-next-line @typescript-eslint/naming-convention -- This is a constant object.
-export const DISPUTE_HIGH_IMPACT_FIELDS: Record<
-	DisputeReason,
-	Record< ProductType, string[] >
-> = {
-	credit_not_processed: {
-		// `customer_signature` (signed delivery proof) is scoped to
-		// physical_product because these disputes commonly take the shape
-		// "I returned the product and never got my refund": proving
-		// delivery corroborates the merchant's defence. The field is
-		// intentionally absent from non-physical cells (no shipping
-		// proof to attach).
-		physical_product: [
-			'customer_signature',
-			'customer_communication',
-			'receipt',
-		],
-		digital_product_or_service: [
-			'customer_communication',
-			'receipt',
-			'refund_refusal_explanation',
-		],
-		offline_service: [
-			'customer_communication',
-			'receipt',
-			'refund_refusal_explanation',
-		],
-		event: [
-			'customer_communication',
-			'receipt',
-			'refund_refusal_explanation',
-		],
-		booking_reservation: [
-			'customer_communication',
-			'receipt',
-			'refund_refusal_explanation',
-		],
-		// `multiple` mirrors `physical_product` as a defensible default
-		// for multi-product orders that may include a physical item.
-		multiple: [ 'customer_signature', 'customer_communication', 'receipt' ],
-		other: [
-			'customer_communication',
-			'receipt',
-			'refund_refusal_explanation',
-		],
-	},
-	duplicate: {
-		physical_product: [
-			'duplicate_charge_explanation',
-			'duplicate_charge_documentation',
-			'shipping_documentation',
-			'receipt',
-		],
-		digital_product_or_service: [
-			'duplicate_charge_explanation',
-			'duplicate_charge_documentation',
-			'receipt',
-		],
-		offline_service: [
-			'duplicate_charge_explanation',
-			'duplicate_charge_documentation',
-			'receipt',
-		],
-		event: [
-			'duplicate_charge_explanation',
-			'duplicate_charge_documentation',
-			'receipt',
-		],
-		booking_reservation: [
-			'duplicate_charge_explanation',
-			'duplicate_charge_documentation',
-			'receipt',
-		],
-		multiple: [
-			'duplicate_charge_explanation',
-			'duplicate_charge_documentation',
-			'shipping_documentation',
-			'receipt',
-		],
-		other: [
-			'duplicate_charge_explanation',
-			'duplicate_charge_documentation',
-			'receipt',
-		],
-	},
-	fraudulent: {
-		// Same lift-based picks across all product types for fraudulent.
-		physical_product: [ 'service_date', 'customer_communication' ],
-		digital_product_or_service: [
-			'service_date',
-			'customer_communication',
-		],
-		offline_service: [ 'service_date', 'customer_communication' ],
-		event: [ 'service_date', 'customer_communication' ],
-		booking_reservation: [ 'service_date', 'customer_communication' ],
-		multiple: [ 'service_date', 'customer_communication' ],
-		other: [ 'service_date', 'customer_communication' ],
-	},
-	general: {
-		physical_product: [ 'receipt', 'customer_communication' ],
-		digital_product_or_service: [ 'receipt', 'customer_communication' ],
-		offline_service: [ 'receipt', 'customer_communication' ],
-		event: [ 'receipt', 'customer_communication' ],
-		booking_reservation: [ 'receipt', 'customer_communication' ],
-		multiple: [ 'receipt', 'customer_communication' ],
-		other: [ 'receipt', 'customer_communication' ],
-	},
-	product_not_received: {
-		physical_product: [
-			'shipping_address',
-			'shipping_tracking_number',
-			'shipping_documentation',
-			'shipping_carrier',
-			'shipping_date',
-		],
-		digital_product_or_service: [
-			'receipt',
-			'customer_communication',
-			'access_activity_log',
-		],
-		offline_service: [ 'receipt', 'customer_communication' ],
-		event: [ 'receipt', 'customer_communication' ],
-		booking_reservation: [ 'receipt', 'customer_communication' ],
-		multiple: [
-			'shipping_address',
-			'shipping_tracking_number',
-			'shipping_documentation',
-			'shipping_carrier',
-			'shipping_date',
-		],
-		other: [ 'receipt', 'customer_communication' ],
-	},
-	product_unacceptable: {
-		physical_product: [
-			'customer_communication',
-			'refund_refusal_explanation',
-			'shipping_documentation',
-		],
-		digital_product_or_service: [
-			'access_activity_log',
-			'customer_communication',
-			'refund_refusal_explanation',
-		],
-		offline_service: [
-			'customer_communication',
-			'refund_refusal_explanation',
-		],
-		// Recommendations for these cells are topical-only (refund_policy,
-		// event/booking documentation). They surface as `optional_missing`
-		// via the existing matrix; no `expected_missing` markers without
-		// data-backed lift.
-		event: [],
-		booking_reservation: [],
-		multiple: [
-			'customer_communication',
-			'refund_refusal_explanation',
-			'shipping_documentation',
-		],
-		other: [],
-	},
-	subscription_canceled: {
-		physical_product: [
-			'cancellation_policy_disclosure',
-			'cancellation_policy',
-			'cancellation_rebuttal',
-		],
-		digital_product_or_service: [
-			'cancellation_policy_disclosure',
-			'cancellation_policy',
-			'cancellation_rebuttal',
-		],
-		offline_service: [
-			'cancellation_policy_disclosure',
-			'cancellation_policy',
-			'cancellation_rebuttal',
-		],
-		event: [
-			'cancellation_policy_disclosure',
-			'cancellation_policy',
-			'cancellation_rebuttal',
-		],
-		booking_reservation: [
-			'cancellation_policy_disclosure',
-			'cancellation_policy',
-			'cancellation_rebuttal',
-		],
-		multiple: [
-			'cancellation_policy_disclosure',
-			'cancellation_policy',
-			'cancellation_rebuttal',
-		],
-		other: [
-			'cancellation_policy_disclosure',
-			'cancellation_policy',
-			'cancellation_rebuttal',
-		],
-	},
-	// No data-backed signal yet: tri-state renders no `expected_missing`
-	// rows for any product type under these reasons.
-	bank_cannot_process: emptyByProductType(),
-	check_returned: emptyByProductType(),
-	customer_initiated: emptyByProductType(),
-	debit_not_authorized: emptyByProductType(),
-	incorrect_account_details: emptyByProductType(),
-	insufficient_funds: emptyByProductType(),
-	noncompliant: emptyByProductType(),
-	unrecognized: emptyByProductType(),
-};
-
-/**
- * Fields that are topically relevant to a dispute reason × product type
- * but lack a defensible win-rate lift signal — and that the wizard's
- * `evidenceMatrix` deliberately omits for that cell. Surfaced by the
- * Dispute Outcome View tri-state renderer as `optional_missing` (muted),
- * never as `expected_missing` (red).
- *
- * This map exists because the wizard matrix and the Outcome View serve
- * different surfaces (pre-response challenge UI vs. post-resolution
- * coaching). Topical recommendations from the post-resolution surface
- * should not bleed into the pre-response wizard, so they live here
- * instead of in `evidenceMatrix`.
- *
- * When a Q-refresh promotes a topical field to a defensible lift signal
- * (≥ +3pp), move the entry from `DISPUTE_TOPICAL_FIELDS` to
- * `DISPUTE_HIGH_IMPACT_FIELDS`. Same shape; pure data move.
- */
-// eslint-disable-next-line @typescript-eslint/naming-convention -- This is a constant object.
-export const DISPUTE_TOPICAL_FIELDS: Record<
-	DisputeReason,
-	Record< ProductType, string[] >
-> = {
-	subscription_canceled: {
-		// physical_product, multiple: refund_policy is already in the
-		// wizard matrix for these cells; no need to duplicate.
-		physical_product: [],
-		digital_product_or_service: [ 'refund_policy' ],
-		offline_service: [ 'refund_policy' ],
-		event: [ 'refund_policy' ],
-		booking_reservation: [ 'refund_policy' ],
-		multiple: [],
-		other: [ 'refund_policy' ],
-	},
-	product_unacceptable: {
-		// physical, digital, offline, event, booking: refund_policy is
-		// already in the wizard matrix for these cells.
-		physical_product: [],
-		digital_product_or_service: [],
-		offline_service: [],
-		event: [],
-		booking_reservation: [],
-		multiple: [],
-		other: [ 'refund_policy' ],
-	},
-	// All other reasons: topicals are already covered by the wizard
-	// matrix for the relevant product types, so no entries are needed
-	// here.
-	bank_cannot_process: emptyByProductType(),
-	check_returned: emptyByProductType(),
-	credit_not_processed: emptyByProductType(),
-	customer_initiated: emptyByProductType(),
-	debit_not_authorized: emptyByProductType(),
-	duplicate: emptyByProductType(),
-	fraudulent: emptyByProductType(),
-	general: emptyByProductType(),
-	incorrect_account_details: emptyByProductType(),
-	insufficient_funds: emptyByProductType(),
-	noncompliant: emptyByProductType(),
-	product_not_received: emptyByProductType(),
-	unrecognized: emptyByProductType(),
-};
-
-/**
- * Human-readable labels used as a fallback when `findMatrixLabel` cannot
- * resolve a label for a `dispute.evidence` key — either because the key
- * is a text/base field not stored in `evidenceMatrix` (which only carries
- * document-upload cells), or because multiple matched wizard cells
- * disagree on the label (collision) and no single status-specific label
- * is appropriate in the post-resolution view.
- *
- * Consumed by `resolveFieldLabel` (called from `getExpectedFieldStatus`).
- */
-// eslint-disable-next-line @typescript-eslint/naming-convention -- This is a constant object.
-const FALLBACK_EVIDENCE_FIELD_LABELS: Record< string, string > = {
-	billing_address: __( 'Billing address', 'woocommerce-payments' ),
-	cancellation_policy_disclosure: __(
-		'Cancellation policy disclosure',
-		'woocommerce-payments'
-	),
-	cancellation_rebuttal: __( 'Cancellation logs', 'woocommerce-payments' ),
-	// Base fields are auto-merged into wizard cells at runtime by
-	// `getRecommendedDocumentFields`, not stored in `evidenceMatrix`. The
-	// outcome-view tri-state helper reads `evidenceMatrix` directly, so it
-	// would otherwise render the raw key. Match the labels the wizard uses.
-	customer_communication: __(
-		'Customer communication',
-		'woocommerce-payments'
-	),
-	customer_signature: __( "Customer's signature", 'woocommerce-payments' ),
-	// Document-upload fields used as high-impact picks. The wizard matrix
-	// labels each per (reason, productType) context (e.g.
-	// `shipping_documentation` is "Return tracking" in CNP cells), but
-	// the outcome view uses neutral, Stripe-aligned labels for cells
-	// where the wizard matrix has no entry to borrow from (notably the
-	// synthesised `multiple` product type).
-	duplicate_charge_documentation: __(
-		'Duplicate charge documentation',
-		'woocommerce-payments'
-	),
-	duplicate_charge_explanation: __(
-		'Duplicate charge explanation',
-		'woocommerce-payments'
-	),
-	product_description: __( 'Product description', 'woocommerce-payments' ),
-	receipt: __( 'Order receipt', 'woocommerce-payments' ),
-	refund_policy: __( 'Refund policy', 'woocommerce-payments' ),
-	refund_refusal_explanation: __(
-		'Refund refusal explanation',
-		'woocommerce-payments'
-	),
-	service_date: __( 'Service date', 'woocommerce-payments' ),
-	shipping_address: __( 'Shipping address', 'woocommerce-payments' ),
-	shipping_carrier: __( 'Shipping carrier', 'woocommerce-payments' ),
-	shipping_date: __( 'Shipping date', 'woocommerce-payments' ),
-	shipping_documentation: __(
-		'Shipping documentation',
-		'woocommerce-payments'
-	),
-	shipping_tracking_number: __(
-		'Shipping tracking number',
-		'woocommerce-payments'
-	),
-	// Catch-all field. Surfaces via the wizard matrix scan (not the
-	// high-impact or topical maps). Wizard cells label it differently per
-	// status branch in some composite-key cells (e.g. "Other documents"
-	// vs "Proof of acceptance" in CNP physical), so the neutral fallback
-	// kicks in via collision detection in `findMatrixLabel`.
-	uncategorized_file: __( 'Other documents', 'woocommerce-payments' ),
-};
-
-/**
- * Find a label for `key` in the wizard matrix, scoped to the cells that
- * apply to the given product type. Composite-key reasons (`duplicate`,
- * `credit_not_processed`) store cells keyed `${productType}__${status}`;
- * we match any cell whose key equals `productType` or starts with
- * `${productType}__`.
- *
- * Collision handling: if matched cells disagree on the label for `key`
- * (the wizard intentionally uses status-specific labels in some composite
- * cells, e.g. `duplicate_charge_documentation` is "Refund receipt" under
- * `__is_duplicate` and "Any additional receipts" under `__is_not_duplicate`),
- * we return undefined so the caller (`resolveFieldLabel`) can fall through
- * to a neutral label from `FALLBACK_EVIDENCE_FIELD_LABELS`. Picking either
- * status-specific label would be misleading in the post-resolution view,
- * which has no wizard-time status to disambiguate.
- *
- * Single-match wins. Multi-match with all-equal labels also wins. Multi-
- * match with disagreement falls through.
- */
-const findMatrixLabel = (
-	reason: string,
-	productType: string,
-	key: string
-): string | undefined => {
-	const productTypeEntries = evidenceMatrix[ reason ];
-	if ( ! productTypeEntries ) {
-		return undefined;
-	}
-
-	const productTypePrefix = `${ productType }__`;
-	const labels = new Set< string >();
-	for ( const [ matrixKey, docs ] of Object.entries( productTypeEntries ) ) {
-		if (
-			matrixKey !== productType &&
-			! matrixKey.startsWith( productTypePrefix )
-		) {
-			continue;
-		}
-		for ( const doc of docs ) {
-			if ( doc.key === key ) {
-				labels.add( doc.label );
-			}
-		}
-	}
-
-	return labels.size === 1 ? [ ...labels ][ 0 ] : undefined;
-};
-
-const resolveFieldLabel = (
-	reason: string,
-	productType: string,
-	key: string
-): string =>
-	findMatrixLabel( reason, productType, key ) ??
-	FALLBACK_EVIDENCE_FIELD_LABELS[ key ] ??
-	key;
-
-const hasMeaningfulValue = ( value: unknown ): boolean => {
-	if ( value === undefined || value === null ) {
-		return false;
-	}
-	if ( typeof value === 'string' ) {
-		return value.trim().length > 0;
-	}
-	if ( typeof value === 'object' ) {
-		return Object.values( value as Record< string, unknown > ).some(
-			hasMeaningfulValue
-		);
-	}
-	return Boolean( value );
-};
-
-const isFieldProvided = (
-	evidence: Record< string, unknown >,
-	key: string
-): boolean => hasMeaningfulValue( evidence[ key ] );
-
-/**
- * Collect every matrix key that applies to the given product type for the
- * reason. Composite-key reasons (`duplicate`, `credit_not_processed`) store
- * cells keyed `${productType}__${status}`; we union every cell whose key
- * starts with `${productType}__` so the optional-missing pool covers all
- * status branches the resolved dispute might have come from.
- *
- * Mirrors the base-field merge that `getRecommendedDocumentFields` applies:
- * when at least one cell matches, `customer_communication` is implicitly
- * recommended (the wizard adds it as a base field for cells that don't
- * already include it). Without this, cells that omit `customer_communication`
- * explicitly (37 of 49 today) would silently drop it from the optional-
- * missing pool in the outcome view.
- */
-const collectMatrixKeys = (
-	reason: string,
-	productType: string
-): Set< string > => {
-	const keys = new Set< string >();
-	const productTypeEntries = evidenceMatrix[ reason ];
-	if ( ! productTypeEntries ) {
-		return keys;
-	}
-
-	const productTypePrefix = `${ productType }__`;
-	let matched = false;
-	for ( const [ matrixKey, docs ] of Object.entries( productTypeEntries ) ) {
-		if (
-			matrixKey !== productType &&
-			! matrixKey.startsWith( productTypePrefix )
-		) {
-			continue;
-		}
-		matched = true;
-		for ( const doc of docs ) {
-			keys.add( doc.key );
-		}
-	}
-	if ( matched ) {
-		keys.add( 'customer_communication' );
-	}
-	return keys;
-};
-
-/**
- * Determine the tri-state status of evidence fields for a (reason, product
- * type) pair.
- *
- * The helper returns one entry per key in the union of three sources:
- *   - `DISPUTE_HIGH_IMPACT_FIELDS[reason][productType]` (red ✗ when missing)
- *   - `DISPUTE_TOPICAL_FIELDS[reason][productType]` (muted — when missing)
- *   - Every field in `evidenceMatrix[reason]` whose cell applies to
- *     `productType`, including composite `${productType}__${status}`
- *     cells (muted — when missing)
- *
- * States:
- *   - `provided`:         `evidence[key]` is a non-empty string (after
- *                         trimming) or an object containing at least one
- *                         non-empty leaf value
- *   - `expected_missing`: key is in the high-impact list for this cell and empty
- *   - `optional_missing`: key is topical or matrix-only, and empty
- *
- * Cells with an empty high-impact list produce no `expected_missing` rows.
- * Unrecognised reason or product type strings return an empty array.
- */
-export const getExpectedFieldStatus = (
-	reason: string,
-	productType: string,
-	evidence: Record< string, unknown >
-): EvidenceFieldStatus[] => {
-	// `emitted` provides cross-source deduplication so a key shared
-	// between sources (e.g., a field that's both high-impact and topical,
-	// or both high-impact and present in the wizard matrix) renders
-	// exactly once, in source-priority order: high-impact, topical,
-	// matrix. Within a single source: high-impact entries are emitted
-	// without a self-dedupe check, so a duplicate in `DISPUTE_HIGH_IMPACT_FIELDS`
-	// surfaces as a duplicate row rather than being silently masked
-	// (we want data bugs in the seed to fail loud). Topical entries and
-	// matrix entries cannot duplicate within their source: topical loops
-	// against `emitted` (so an in-source duplicate is masked), and
-	// `matrixKeys` is a `Set` by construction.
-	const highImpactKeys =
-		DISPUTE_HIGH_IMPACT_FIELDS[ reason as DisputeReason ]?.[
-			productType as ProductType
-		] ?? [];
-
-	const topicalKeys =
-		DISPUTE_TOPICAL_FIELDS[ reason as DisputeReason ]?.[
-			productType as ProductType
-		] ?? [];
-
-	const matrixKeys = collectMatrixKeys( reason, productType );
-
-	const result: EvidenceFieldStatus[] = [];
-	const emitted = new Set< string >();
-
-	for ( const key of highImpactKeys ) {
-		result.push( {
-			key,
-			label: resolveFieldLabel( reason, productType, key ),
-			state: isFieldProvided( evidence, key )
-				? 'provided'
-				: 'expected_missing',
-		} );
-		emitted.add( key );
-	}
-
-	for ( const key of topicalKeys ) {
-		if ( emitted.has( key ) ) {
-			continue;
-		}
-		result.push( {
-			key,
-			label: resolveFieldLabel( reason, productType, key ),
-			state: isFieldProvided( evidence, key )
-				? 'provided'
-				: 'optional_missing',
-		} );
-		emitted.add( key );
-	}
-
-	for ( const key of matrixKeys ) {
-		if ( emitted.has( key ) ) {
-			continue;
-		}
-		result.push( {
-			key,
-			label: resolveFieldLabel( reason, productType, key ),
-			state: isFieldProvided( evidence, key )
-				? 'provided'
-				: 'optional_missing',
-		} );
-		emitted.add( key );
-	}
-
-	return result;
 };
